% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRRAT-package.R
\docType{package}
\name{BRRAT-package}
\alias{BRRAT-package}
\title{The 'BRRAT' package.}
\description{
Function to apply the Bayesian Regression Robustness Assessment Test,
to evaluate if model simulations are independent of the climate input used,
suggesting a robust model.

This package makes use of Stan to fit the linear regression model,
and the precompiled model created for the package with rstantools.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org

Gibbs et al. (2025)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-s-gibbs/SWTools}
}

}
\author{
\strong{Maintainer}: Matt Gibbs \email{matt.gibbs@csiro.au}

}
