% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRRAT-package.R
\docType{_PACKAGE}
\name{BRRAT-package}
\alias{BRRAT-package}
\alias{BRRAT}
\title{The 'BRRAT' package.}
\description{
A DESCRIPTION OF THE PACKAGE. Package generated using rstantools
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
