% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRRAT.R
\name{BRRAT}
\alias{BRRAT}
\title{Bayesian Regression Robustness Assessment Test}
\usage{
BRRAT(Sim, Obs, Clim, ...)
}
\arguments{
\item{Sim}{Annual time series of simulated values (typically streamflow volume)}

\item{Obs}{Annual time series of observed values}

\item{Clim}{Annual time series of independent climate values}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
list with slopes for Sim (beta1), Sim2 if provided (beta2), and rstan model fit object for testing of convergence etc, lambda exponent used for BC transformation
}
\description{
Bayesian Regression Robustness Assessment Test
}
\examples{
samples <- 30
Clim <- sort(pmax(0,rnorm(samples,600,100))) #generate some annual rainfall totals
Obs <- pmax(0,(Clim - 150) - 450 * tanh((Clim - 150)/450)) #calculate annual runoff from tanh curve
Sim1 <- Obs*seq(1,0.7,length.out=length(Obs))*rnorm(samples,1,0.05) #non-robust model, underestimate the wetter years
Sim2 <- Obs*rnorm(samples,1,0.05) #robust model, only random error
slopes1 <- BRRAT(Sim1,Obs,Clim)
slopes2 <- BRRAT(Sim2,Obs,Clim)

dat <- data.frame(beta = c(slopes1$beta, slopes2$beta),
                  id = c(rep("Sim1",length(slopes1$beta)),
                  rep("Sim2",length(slopes2$beta))))
ggplot2::ggplot(dat)+geom_density(aes(beta,fill=id))



}
