% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRRAT.R
\name{BRRAT}
\alias{BRRAT}
\title{Bayesian Regression Robustness Assessment Test}
\usage{
BRRAT(Sim, Obs, ...)
}
\arguments{
\item{Sim}{Time series of simulated values (typically annual streamflow volume)}

\item{Obs}{Time series of observed values}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains, cores).}
}
\value{
a list with members:
\describe{
\item{beta}{samples of slope values from the posterior distribution}
\item{fit}{rstan model  object for testing of convergence etc}
\item{lambda}{value of exponent used for BC transformation}
}
}
\description{
Function to apply the BRRAT. The input \code{Sim} and \code{Obs} time series are first
transformed to near normal distributions using a boxcox transformation, with
the lambda exponent fit to the \code{obs} data.
}
\details{
A linear regression is fit between the error between the transformed \code{Sim}
and \code{Obs} data and the corresponding \code{Clim} data, to test for a dependent
relationship between the error and climate, indicating a model that is not
robust to the input climate data. Stan is used to fit the linear regression,
and hence a distribution of slope values are returned.
}
\examples{

\dontrun{
#generate 2 models of annual streamflow
samples <- 50
#generate some annual rainfall totals
Clim <- sort(pmax(0,rnorm(samples,600,100)))
#calculate annual runoff from tanh curve
observed <- pmax(0,(Clim - 150) - 450 * tanh((Clim - 150)/450))
Obs <- data.frame(P=Clim,Q=observed,year=1:length(Clim))

#non-robust model, underestimate the wetter years
Sim1 <- observed*seq(1,0.7,length.out=length(observed))*rnorm(samples,1,0.05)
Sim1 <- data.frame(Q=Sim1,year=1:length(Clim))
#robust model, only random error
Sim2 <- observed*rnorm(samples,1,0.05)
Sim2 <- data.frame(Q=Sim2,year=1:length(Clim))

#apply BRRAT test to both, reduce chains and iters to speed up example.
fit1 <- BRRAT(Sim1,Obs,chains=1,iter=500)
fit2 <- BRRAT(Sim2,Obs,chains=1,iter=500)

dat <- data.frame(b = c(fit1$beta_draws,fit2$beta_draws),
id = c(rep("fit1",length(fit1$beta_draws)),rep("fit2",length(fit2$beta_draws)))
ggplot(dat)+geom_density(aes(b,fill=id)) #fit2 is centered on 0, fit1 has a negative slope
#can also be viewed using the plotting function
plot_BRRAT_population(fit1)
plot_BRRAT_population(fit2)

#Summarise statistics of the mean can be viewed using
summarise_BRRAT(fit1)
summarise_BRRAT(fit2)
}

}
